pyproject.toml
[project]
name = "invest-notify"
version = "0.1.0"
description = "Track Taiwan stock prices periodically and plot 3-month trends"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "apscheduler>=3.10.4",
  "matplotlib>=3.8.0",
  "pandas>=2.2.0",
  "pydantic>=2.6.0",
  "pyyaml>=6.0.1",
  "requests>=2.31.0",
  "seaborn>=0.13.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.3.0",
]

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/invest_notify"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
